feat(shared): centralise provider credential TypedDicts in harnessiq/shared/credentials.py and fix config merge artifacts
