+++ harnessiq/providers/phantombuster/credentials.py
@@ -1,17 +1,5 @@
-"""PhantomBuster credential configuration type."""
+"""PhantomBuster credential configuration type — re-exported from shared."""
 
-from __future__ import annotations
-
-from harnessiq.config.models import ProviderCredentialConfig
-
-
-class PhantomBusterCredentials(ProviderCredentialConfig, total=True):
-    """Credentials for authenticating with the PhantomBuster API.
-
-    Pass the *api_key* in the ``X-Phantombuster-Key`` request header.
-    """
-
-    api_key: str
-
+from harnessiq.shared.credentials import PhantomBusterCredentials as PhantomBusterCredentials
 
 __all__ = ["PhantomBusterCredentials"]
