+++ harnessiq/providers/proxycurl/credentials.py
@@ -1,20 +1,5 @@
-"""Proxycurl credential configuration type."""
+"""Proxycurl credential configuration type — re-exported from shared."""
 
-from __future__ import annotations
-
-from harnessiq.config.models import ProviderCredentialConfig
-
-
-class ProxycurlCredentials(ProviderCredentialConfig, total=True):
-    """Credentials for authenticating with the Proxycurl API.
-
-    NOTE: Proxycurl shut down in January 2025 following a LinkedIn lawsuit.
-    This provider is preserved for reference only.
-
-    Pass the *api_key* as a Bearer token in the ``Authorization`` request header.
-    """
-
-    api_key: str
-
+from harnessiq.shared.credentials import ProxycurlCredentials as ProxycurlCredentials
 
 __all__ = ["ProxycurlCredentials"]
